♻️(core) remove dependency to factory in production
The helpers.py file which is used in production was depending on
factory because of 2 functions:
- the `create_text_plugin` function is actually only used in tests,
  let's move it to factories.py,
- the `create_i18n_page` function was using faker to generate random
  title when the title was not passed. Let's make title a required
  field and let the page extension factories generate a random title
  if they need one (this will soon be refactored to make use of our
  new PageFactory).
