+++ tests/apps/courses/test_models_organization.py
@@ -149,7 +149,7 @@ class OrganizationModelsTestCase(TestCase):
         # which is there to exclude snapshots but also acts on the main course page and
         # checks its parent (so the root page) and the duplicate comes from the fact that
         # the parent has a draft and a public page... so "cms_pages" has a cardinality of 2
-        root_page = create_i18n_page(published=True)
+        root_page = create_i18n_page("my title", published=True)
 
         organization = OrganizationFactory(should_publish=True)
         course = CourseFactory(
